@@ -13,45 +13,45 @@
   },
   "private": true,
   "dependencies": {
-    "@angular-devkit/core": "^19.2.11",
-    "@angular/animations": "^19.2.10",
-    "@angular/cdk": "^19.2.15",
-    "@angular/common": "^19.2.10",
-    "@angular/compiler": "^19.2.10",
-    "@angular/core": "^19.2.10",
-    "@angular/forms": "^19.2.10",
-    "@angular/material": "^19.2.15",
-    "@angular/platform-browser": "^19.2.10",
-    "@angular/platform-browser-dynamic": "^19.2.10",
-    "@angular/router": "^19.2.10",
-    "eslint": "^9.20.1",
+    "@angular-devkit/core": "^20.3.14",
+    "@angular/animations": "^20.3.16",
+    "@angular/cdk": "^20.2.14",
+    "@angular/common": "^20.3.16",
+    "@angular/compiler": "^20.3.16",
+    "@angular/core": "^20.3.16",
+    "@angular/forms": "^20.3.16",
+    "@angular/material": "^20.2.14",
+    "@angular/platform-browser": "^20.3.16",
+    "@angular/platform-browser-dynamic": "^20.3.16",
+    "@angular/router": "^20.3.16",
+    "eslint": "^9.28.0",
     "ngxtension": "^4.3.2",
     "rxjs": "^7.8.1",
     "tslib": "^2.8.1",
     "zone.js": "^0.15.0"
   },
   "devDependencies": {
-    "@angular-devkit/architect": "^0.1902.11",
-    "@angular-devkit/build-angular": "^19.2.11",
-    "@angular-eslint/builder": "~19.4.0",
-    "@angular-eslint/eslint-plugin": "~19.4.0",
-    "@angular-eslint/eslint-plugin-template": "~19.4.0",
-    "@angular-eslint/schematics": "~19.4.0",
-    "@angular-eslint/template-parser": "~19.4.0",
-    "@angular/cli": "^19.2.11",
-    "@angular/compiler-cli": "^19.2.10",
-    "@angular/language-service": "^19.2.10",
+    "@angular-devkit/architect": "^0.2003.14",
+    "@angular-devkit/build-angular": "^20.3.14",
+    "@angular-eslint/builder": "~20.7.0",
+    "@angular-eslint/eslint-plugin": "~20.7.0",
+    "@angular-eslint/eslint-plugin-template": "~20.7.0",
+    "@angular-eslint/schematics": "~20.7.0",
+    "@angular-eslint/template-parser": "~20.7.0",
+    "@angular/cli": "^20.3.14",
+    "@angular/compiler-cli": "^20.3.16",
+    "@angular/language-service": "^20.3.16",
     "@cypress/schematic": "^3.0.0",
     "@eslint/eslintrc": "^3.3.1",
     "@eslint/js": "^9.24.0",
     "@floogulinc/cypress-mongo-seeder": "^1.1.1",
-    "@types/jasmine": "^5.1.6",
-    "@types/node": "^22.13.4",
+    "@types/jasmine": "^6.0.0",
+    "@types/node": "^22.13.17",
     "@typescript-eslint/eslint-plugin": "^8.24.1",
-    "@typescript-eslint/parser": "^8.24.1",
-    "cypress": "^14.0.3",
+    "@typescript-eslint/parser": "^8.54.0",
+    "cypress": "^15.10.0",
     "eslint-plugin-import": "2.31.0",
-    "eslint-plugin-jsdoc": "^50.6.9",
+    "eslint-plugin-jsdoc": "^62.5.2",
     "eslint-plugin-prefer-arrow": "1.2.3",
     "jasmine-core": "^5.6.0",
     "jasmine-spec-reporter": "~7.0.0",
@@ -61,6 +61,6 @@
     "karma-jasmine": "~5.1.0",
     "karma-jasmine-html-reporter": "^2.1.0",
     "ts-node": "^10.9.2",
-    "typescript": "^5.7.3"
+    "typescript": "^5.9.3"
   }
 }
