@@ -1,4 +1,4 @@
-import { Component, Signal } from '@angular/core';
+import { Component, Signal, inject } from '@angular/core';
 import { toSignal } from '@angular/core/rxjs-interop';
 import { CompanyCardComponent } from '../company-card/company-card.component';
 import { UserService } from '../users/user.service';
@@ -11,7 +11,7 @@ import { Company } from './company';
   styleUrl: './company-list.component.scss'
 })
 export class CompanyListComponent {
-  companies: Signal<Company[]> = toSignal(this.userService.getCompanies());
+  private userService = inject(UserService);
 
-  constructor(private userService: UserService) {}
+  companies: Signal<Company[]> = toSignal(this.userService.getCompanies());
 }
