@@ -1,6 +1,4 @@
 import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
-import { MatCardModule } from '@angular/material/card';
-import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { UserCardComponent } from './user-card.component';
 import { User } from './user';
 
@@ -12,8 +10,6 @@ describe('UserCardComponent', () => {
   beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
       imports: [
-        BrowserAnimationsModule,
-        MatCardModule,
         UserCardComponent
       ]
     })
@@ -23,17 +19,15 @@ describe('UserCardComponent', () => {
   beforeEach(() => {
     fixture = TestBed.createComponent(UserCardComponent);
     component = fixture.componentInstance;
-    TestBed.runInInjectionContext(() => {
-      expectedUser = {
-        _id: 'chris_id',
-        name: 'Chris',
-        age: 25,
-        company: 'UMM',
-        email: 'chris@this.that',
-        role: 'admin',
-        avatar: 'https://gravatar.com/avatar/8c9616d6cc5de638ea6920fb5d65fc6c?d=identicon'
-      };
-    });
+    expectedUser = {
+      _id: 'chris_id',
+      name: 'Chris',
+      age: 25,
+      company: 'UMM',
+      email: 'chris@this.that',
+      role: 'admin',
+      avatar: 'https://gravatar.com/avatar/8c9616d6cc5de638ea6920fb5d65fc6c?d=identicon'
+    };
     fixture.componentRef.setInput('user', expectedUser);
     fixture.detectChanges();
   });
